Create Chef in Vaccination Queue
There are 
�
N people in the vaccination queue, Chef is standing on the 
�
�
ℎ
P 
th
  position from the front of the queue. It takes 
�
X minutes to vaccinate a child and 
�
Y minutes to vaccinate an elderly person. Assume Chef is an elderly person.

You are given a binary array 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
  of length 
�
N, where 
�
�
=
1
A 
i
​
 =1 denotes there is an elderly person standing on the 
�
�
ℎ
i 
th
  position of the queue, 
�
�
=
0
A 
i
​
 =0 denotes there is a child standing on the 
�
�
ℎ
i 
th
  position of the queue. You are also given the three integers 
�
,
�
,
�
P,X,Y. Find the number of minutes after which Chef's vaccination will be completed.