Create Elections in Chefland
There are 
101
101 citizens in Chefland. It is election time in Chefland and 
3
3 parties, 
�
,
�
A,B, and 
�
C are contesting the elections. Party 
�
A receives 
�
�
X 
A
​
  votes, party 
�
B receives 
�
�
X 
B
​
  votes, and party 
�
C receives 
�
�
X 
C
​
  votes.

The constitution of Chefland requires a particular party to receive a clear majority to form the government. A party is said to have a clear majority if it receives strictly greater than 
50
50 votes.

If any party has a clear majority, print the winning party (A, B or C). Otherwise, print NOTA.