Create Debug this code - Determine the Score
The code given in the IDE is incorrect - Try and debug this program!!!

There is a problem worth 
�
X points and it has 
10
10 test cases.
It is known that 
�
X is a multiple of 
10
10.
It is known that each test case is worth the same number of points.
You pass 
�
N test cases among them. Determine the score that you will get.

Input Format
First line will contain 
�
T, number of test cases. Then the test cases follow.
Each test case contains of a single line of input, two integers 
�
X and 
�
N, the total points for the problem and the number of test cases which pass.
Output Format
For each test case, output the points scored by you