Create Max minus Min
Given 3 integers - 
�
A, 
�
B and 
�
C - you need to find the difference between the highest and the lowest of the given 3 integers.

Input Format
The first line of input will contain a single integer 
�
T, denoting the number of test cases.
Each test case consists of 
3
3 integers 
�
,
�
,
�
A,B,C.
Output Format
For each test case, output on a new line the difference between the highest and the lowest of the given 3 integers.

Hint - We know Math.min() can be used to check minimum of 2 numbers only, but, there's a trick to use the Math.min() method to check min of 3 numbers as well, try to figure it out. Same goes to finding maximum.