Create Dominant Element
You are given an array 
�
A of length 
�
N. An element 
�
X is said to be dominant if the frequency of 
�
X in 
�
A is strictly greater than the frequency of any other element in the 
�
A.

For example, if 
�
=
[
2
,
1
,
4
,
4
,
4
]
A=[2,1,4,4,4] then 
4
4 is a dominant element since its frequency is higher than the frequency of any other element in 
�
A.

Find if there exists any dominant element in 
�
A.

Input Format
The first line of input contains a single integer 
�
T — the number of test cases. Then the test cases follow.
The first line of each test case contains an integer 
�
N — the size of the array 
�
A.
The second line of each test case contains 
�
N space-separated integers 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
  denoting the array 
�
A.
Output Format
For each test case, output YES if there exists any dominant element in 
�
A. Otherwise, output NO.

You may print each character of YES and NO in uppercase or lowercase (for example, yes, yEs, Yes will be considered identical).