Create  Luigi and Uniformity
Luigi has an array 
�
A of 
�
N positive integers. He wants to make all elements of the array equal.

In one move, he can:

Choose an index 
�
i 
(
1
≤
�
≤
�
)
(1≤i≤N) and divide the element 
�
�
A 
i
​
  by any one of its divisors.
In other words, he can choose a positive integer 
�
X such that 
�
∣
�
�
X∣A 
i
​
  and set 
�
�
:
=
�
�
�
A 
i
​
 := 
X
A 
i
​
 
​
 .
Find the minimum number of moves required to make all the elements of the array equal.

Input Format
The first line of input will contain a single integer 
�
T, denoting the number of test cases.
Each test case consists of two lines of input.
The first line of each test case contains 
�
N, the size of array 
�
A.
The second line of each test case contains 
�
N space-separated integers, the elements of array 
�
A.
Output Format
For each test case, output on a new line, the minimum number of moves required to make all elements of the array equal.