Create Strong Elements
Chef has an array 
�
A of length 
�
N.

An index 
�
i is called strong if we can change the gcd of the whole array just by changing the value of 
�
�
A 
i
​
 .

Determine the number of strong indices in the array.

Input Format
First line will contain 
�
T, number of test cases. Then the test cases follow.
First line of each test case contains an integer 
�
N denoting the size of the array 
�
A.
Second line contains 
�
N space separated integers 
�
1
,
�
2
,
…
,
�
�
A 
1
​
 ,A 
2
​
 ,…,A 
N
​
  - denoting the array 
�
A.
Output Format
For each test case, output the number of strong indices in the array.